Img, vid and audio flows working
