+++ app/src/stores/nugget.js
@@ -3,31 +3,35 @@ import { defineStore } from "pinia";
 import { useStorage } from "@vueuse/core";
 
 import { useNuggets } from "@/composables/Nuggets";
+import { useMulticorder } from "@/composables/Multicorder";
+
+const {
+  tmpStore,
+
+  // VIDEO / IMAGES
+  showCamera,
+  showCameraDialog,
+  saveVideoSource,
+  videoSource,
+  selectedVideoDevice,
+  preferredCamera,
+  saveVideoChunk,
+  tmpImages,
+  tmpVideos,
+
+  //AUDIO
+  tmpAudios,
+} = useMulticorder();
 
 const {
   addNuggetAttachments,
-    addNuggetAsset,
-    createNugget,
-    getNugget,
-    getNuggetAssets,
-    newFileTimestamp,
-    readOPFSFile,
-    setGeoLocation,
-    tmpStore,
-
-    // VIDEO / IMAGES
-    showCamera,
-    showCameraDialog,
-    saveVideoSource,
-    videoSource,
-    selectedVideoDevice,
-    preferredCamera,
-    saveVideoChunk,
-    tmpImages,
-    tmpVideos,
-
-    //AUDIO
-    tmpAudios,
+  addNuggetAsset,
+  createNugget,
+  getNugget,
+  getNuggetAssets,
+  newFileTimestamp,
+  readOPFSFile,
+  setGeoLocation,
 } = useNuggets();
 
 export const useNuggetStore = defineStore("nugget", () => {
@@ -43,41 +47,42 @@ export const useNuggetStore = defineStore("nugget", () => {
 
 
   // ACTIONS / FUNCTIONS
-  const startNuggetExport = async(nuggetId) => {
+  const startNuggetExport = async (nuggetId) => {
 
-    if(pendingExports.value.has(nuggetId)) {
-      console.log(`Export already in process for nugget ${nuggetId}` )
+    if (pendingExports.value.has(nuggetId)) {
+      console.log(`Export already in process for nugget ${nuggetId}`)
       // return;
     } // else {
-      const nowTime = new Date().toISOString();
-      pendingExports.value.set(nuggetId, { createdAt: nowTime });
-      const exportId = await Nug.startExport(nuggetId);
-      const jobData = { nuggetId: nuggetId, createdAt: nowTime, exportId: exportId };
-      pendingExports.value.set(nuggetId, jobData);
-      return jobData;
-   //}
+    const nowTime = new Date().toISOString();
+    pendingExports.value.set(nuggetId, { createdAt: nowTime });
+    const exportId = await Nug.startExport(nuggetId);
+    const jobData = { nuggetId: nuggetId, createdAt: nowTime, exportId: exportId };
+    pendingExports.value.set(nuggetId, jobData);
+    return jobData;
+    //}
+  }
+
+  const resetMulticorder = () => {
+    mc.$reset();
   }
 
   return {
     // State
 
-    // Getters
+    // MULTICORDER
     tmpImages,
     tmpVideos,
     selectedVideoDevice,
     preferredCamera,
     showCameraDialog,
+    videoSource,
+    tmpAudios,
     saveVideoSource,
     showCamera,
-    videoSource,
-
-    tmpAudios,
+    tmpStore,
 
 
-
-    // Actions/Functions
-
-
+    // NUGGETS
     addNuggetAttachments,
     addNuggetAsset,
     createNugget,
@@ -85,12 +90,7 @@ export const useNuggetStore = defineStore("nugget", () => {
     getNuggetAssets,
     newFileTimestamp,
     readOPFSFile,
-
     setGeoLocation,
-
-    tmpStore,
-
-
     startNuggetExport,
   }
 });
